# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mahzen API
  description: |
    Mahzen is a knowledge management platform that provides CRUD operations for
    entries and tags, keyword and semantic search, and user authentication.

    The HTTP API is served via Gin on port 8080. When TLS is configured, the
    server supports both HTTP/2 and HTTP/3 (QUIC).

    **Authentication**: Requests are authenticated via JWT Bearer tokens in the
    `Authorization` header. The auth middleware validates the token and sets
    `X-User-Id` internally. Public endpoints (register, login, refresh) do not
    require authentication.

    **Visibility**: Entries are either `public` (visible to all) or `private`
    (visible only to the owner). Listing and search endpoints enforce this at the
    database and search-engine level.

    **Pagination**: `limit` defaults to 20 when omitted or <= 0.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development (Gin)
...
paths:
...
  /v1/tags:
    post:
      operationId: CreateTag
      summary: Create a new tag
      tags: [Tags]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "200":
          description: Tag created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagResponse"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: ListTags
      summary: List all tags
      tags: [Tags]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated list of tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTagsResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/tags/{id}:
    get:
      operationId: GetTag
      summary: Get a single tag by ID
      tags: [Tags]
      parameters:
        - $ref: "#/components/parameters/TagID"
      responses:
        "200":
          description: Tag found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: DeleteTag
      summary: Delete a tag
      tags: [Tags]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TagID"
      responses:
        "200":
          description: Tag deleted successfully
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/entries/{entry_id}/tags:
    post:
      operationId: AttachTag
      summary: Attach a tag to an entry
      tags: [Tags]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EntryIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachTagRequest"
      responses:
        "200":
          description: Tag attached successfully
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/entries/{entry_id}/tags/{tag_id}:
    delete:
      operationId: DetachTag
      summary: Detach a tag from an entry
      tags: [Tags]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EntryIDPath"
        - name: tag_id
          in: path
          required: true
          description: ID of the tag to detach
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tag detached successfully
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Search
  # ---------------------------------------------------------------------------
...
  /v1/users/me:
    get:
      operationId: GetCurrentUser
      summary: Get the currently authenticated user
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCurrentUserResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"

# =============================================================================
# Components
# =============================================================================
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    TagID:
      name: id
      in: path
      required: true
      description: Tag UUID
      schema:
        type: string
        format: uuid
    EntryIDPath:
      name: entry_id
      in: path
      required: true
      description: Entry UUID (used in tag association endpoints)
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      required: false
      description: "Maximum number of items to return (default: 20)"
      schema:
        type: integer
        format: int32
        default: 20
        minimum: 1
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
...
  schemas:
...
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
    CreateTagRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    TagResponse:
      type: object
      properties:
        tag:
          $ref: "#/components/schemas/Tag"
    ListTagsResponse:
      type: object
      properties:
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        total:
          type: integer
          format: int32
    AttachTagRequest:
      type: object
      required: [tag_id]
      properties:
        tag_id:
          type: string
          format: uuid
          description: ID of the tag to attach

    # -- Search ----------------------------------------------------------------
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
    GetCurrentUserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"

    # -- Error -----------------------------------------------------------------
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
...
    Unauthenticated:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"