# internal/infra/postgres/query/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package query

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Entry struct {
	ID         pgtype.UUID        `json:"id"`
	UserID     pgtype.UUID        `json:"user_id"`
	Title      string             `json:"title"`
	Content    string             `json:"content"`
	Summary    string             `json:"summary"`
	S3Key      string             `json:"s3_key"`
	Visibility string             `json:"visibility"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `json:"updated_at"`
	Path       string             `json:"path"`
}

type EntryTag struct {
	EntryID pgtype.UUID `json:"entry_id"`
	TagID   pgtype.UUID `json:"tag_id"`
}

type RefreshToken struct {
	ID        pgtype.UUID        `json:"id"`
	UserID    pgtype.UUID        `json:"user_id"`
	TokenHash string             `json:"token_hash"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type Tag struct {
	ID        pgtype.UUID        `json:"id"`
	Name      string             `json:"name"`
	Slug      string             `json:"slug"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type User struct {
	ID           pgtype.UUID        `json:"id"`
	Email        string             `json:"email"`
	DisplayName  string             `json:"display_name"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	PasswordHash string             `json:"password_hash"`
}